Moved GoogleSensorParser to parser.py file.
